Car generator
Created a car generator and seperated class Car from simulator and added a car generator example to simulator
